- Introduce file and memory storage backends for persistence. - Refactor the codes to use abstract class
