+++ examples/nextjs/.env.example
@@ -1,8 +1,14 @@
-# Redis Configuration
-# Required for session storage
+# Storage Configuration
+# Priority: MCP_TS_STORAGE_TYPE > REDIS_URL > MCP_TS_STORAGE_FILE > Memory (Default)
+# MCP_TS_STORAGE_TYPE=redis|file|memory
+
+# 1. Redis Storage (Recommended for production)
 REDIS_URL=your-redis-url
 # Optional: Redis password
 # REDIS_PASSWORD=your-redis-password
+
+# 2. File Storage (Persistent local storage)
+# MCP_TS_STORAGE_FILE=./sessions.json
 
 # Next.js Configuration
 NEXT_PUBLIC_API_URL=http://localhost:3000
